+++ src/frontend/types/yaksa_blkindx.c
@@ -8,7 +8,8 @@
 #include <stdlib.h>
 #include <assert.h>
 
-int yaksi_type_create_hindexed_block(int count, int blocklength, const intptr_t * array_of_displs,
+int yaksi_type_create_hindexed_block(intptr_t count, intptr_t blocklength,
+                                     const intptr_t * array_of_displs,
                                      yaksi_type_s * intype, yaksi_type_s ** newtype)
 {
     int rc = YAKSA_SUCCESS;
@@ -107,7 +108,8 @@ int yaksi_type_create_hindexed_block(int count, int blocklength, const intptr_t 
     goto fn_exit;
 }
 
-int yaksa_type_create_hindexed_block(int count, int blocklength, const intptr_t * array_of_displs,
+int yaksa_type_create_hindexed_block(intptr_t count, intptr_t blocklength,
+                                     const intptr_t * array_of_displs,
                                      yaksa_type_t oldtype, yaksa_info_t info,
                                      yaksa_type_t * newtype)
 {
@@ -137,7 +139,8 @@ int yaksa_type_create_hindexed_block(int count, int blocklength, const intptr_t 
     goto fn_exit;
 }
 
-int yaksa_type_create_indexed_block(int count, int blocklength, const int *array_of_displs,
+int yaksa_type_create_indexed_block(intptr_t count, intptr_t blocklength,
+                                    const intptr_t * array_of_displs,
                                     yaksa_type_t oldtype, yaksa_info_t info, yaksa_type_t * newtype)
 {
     intptr_t *real_array_of_displs = NULL;
@@ -157,7 +160,7 @@ int yaksa_type_create_indexed_block(int count, int blocklength, const int *array
 
     real_array_of_displs = (intptr_t *) malloc(count * sizeof(intptr_t));
 
-    for (int i = 0; i < count; i++)
+    for (intptr_t i = 0; i < count; i++)
         real_array_of_displs[i] = array_of_displs[i] * intype->extent;
 
     yaksi_type_s *outtype;
